+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Change from `search-graphql` to `search-resolver`
+
 ## [0.11.8] - 2020-04-08
 
 ### Fixed